+++ packages/networks/src/networks/vNaga/envs/naga-staging/naga-staging.env.ts
@@ -28,7 +28,6 @@ export class NagaStagingEnvironment extends BaseNetworkEnvironment<
         realmId: DEFAULT_REALM_ID,
       },
       services: {
-        authServiceBaseUrl: 'https://naga-auth-service.getlit.dev',
         loginServiceBaseUrl: 'https://login.litgateway.com',
       },
       minimumThreshold: MINIMUM_THRESHOLD,