+++ packages/networks/src/networks/vNaga/shared/managers/pricing-manager/schema.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod';
 import { getUserMaxPrice } from './getUserMaxPrice';
 import { PRODUCT_IDS } from './constants';
+import { PRODUCT_ID_VALUES } from '@lit-protocol/constants';
 
 export const PricingContextSchema = z
   .object({
@@ -20,11 +21,13 @@ export const PricingContextSchema = z
 
     return {
       product: {
-        id: PRODUCT_IDS[product],
+        id: Number(PRODUCT_IDS[product]) as PRODUCT_ID_VALUES,
         name: product,
       },
       userMaxPrice: _userMaxPrice,
-      nodePrices,
+
+      // This aligns the Zod-inferred type with the function signature, removes the optionality error, and lets you drop the @ts-ignore/cast.
+      nodePrices: nodePrices as { url: string; prices: bigint[] }[],
       threshold,
     };
   });