+++ packages/types/src/lib/types.ts
@@ -32,7 +32,7 @@ import type {
 
 export type Optional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 
-export {
+export type {
   DefinedJson,
   DefinedLiteral,
   Json,