+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @lit-protocol/constants
 
+## 8.0.1-test.9
+
+### Patch Changes
+
+- test
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.9
+
+## 8.0.1-test.8
+
+### Patch Changes
+
+- testing
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.8
+
+## 8.0.1-test.7
+
+### Patch Changes
+
+- test
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.7
+
+## 8.0.1-test.6
+
+### Patch Changes
+
+- test
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.6
+
+## 8.0.1-test.5
+
+### Patch Changes
+
+- testing
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.5
+
+## 8.0.1-test.4
+
+### Patch Changes
+
+- testing
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.4
+
+## 8.0.1-test.3
+
+### Patch Changes
+
+- put console logs to debug
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.3
+
+## 8.0.1-test.2
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.2
+
+## 8.0.1-test.1
+
+### Patch Changes
+
+- fix package metadata so bundlers load the CommonJS builds again
+- Updated dependencies
+  - @lit-protocol/contracts@0.5.2-test.1
+
+## 8.0.1-test.0
+
+### Patch Changes
+
+- first pnpm complete release test for naga-test
+- Updated dependencies
+- Updated dependencies [642cd0b]
+  - @lit-protocol/contracts@0.5.2-test.0
+
 ## 8.0.0
 
 ### Major Changes